requirements.txt
Flask==0.10.1
Flask-Login==0.2.7
Jinja2==2.7.1
MarkupSafe==0.18
PyBrowserID==0.9.1
Werkzeug==0.9.4
argparse==1.2.1
bcrypt==1.0.2
cffi==0.7.2
gevent==0.13.8
greenlet==0.4.1
itsdangerous==0.23
pycparser==2.10
requests==2.0.0
wsgiref==0.1.2
